+++ app/src/main/java/hk/edu/hkbu/comp/lab01/json/User.kt
@@ -10,7 +10,7 @@ data class User(
         val level: String,
         val gender: String,
         val status: String,
-        val create_time: Int,
+        val create_time: Long,
         val level_name: String,
         val is_following: Boolean,
         val is_blocked: Boolean,
